refactor: extract ui sidebar helpers
